Moved config_reader and added a server from yaml example.
