+++ examples/basic_server_recipe.yaml
@@ -1,0 +1,121 @@
+# DEV:RW:DOUBLE1:
+#   description: "An example double PV"
+#   type: "DOUBLE"
+#   initial: 3.0
+
+# DEV:RW:DOUBLE2:
+#   description: "An example double PV with units"
+#   type: "DOUBLE"
+#   initial: 5.0
+#   display:
+#     units: 'v'
+
+# DEV:RW:DOUBLE3:
+#   description: "A different default value for the PV"
+#   type: "DOUBLE"
+#   initial: 17.5
+#   control:
+#     low: 1
+#     high: 20
+#   valueAlarm:
+#     low_warning: 2
+#     high_alarm: 9
+
+# DEV:RO:DOUBLE4:
+#   description: "A read-only double"
+#   type: "DOUBLE"
+#   display:
+#     units: "Bq"
+#     format: "EXPONENTIAL"
+#     precision: 3
+#   initial: 17.5
+#   read_only: True
+
+# # DEV:RW:DOUBLE5:
+# #   description: "A double array"
+# #   type: "DOUBLE"
+# #   initial: [3,4,5,17.5,7,8]
+# #   control:
+# #     low: 1
+# #     high: 20
+
+# # DEV:RW:DOUBLEARR2:
+# #   description: "A double array using the default value 0 which is outside the control range"
+# #   type: "DOUBLE"
+# #   array_size: 6
+# #   control:
+# #     low: 1
+# #     high: 20
+
+# DEV:RW:DOUBLE6:
+#   description: "A PV with display limits configured but no precision etc"
+#   type: "DOUBLE"
+#   initial: 6.0
+#   display:
+#     low: -1.0
+#     high: 1.0
+
+# DEV:RW:INT2:
+#   description: "An example integer PV with a forward link"
+#   type: "INTEGER"
+#   initial: 0
+#   #forward_links: "DEV:RW:DOUBLECALC"
+
+DEV:RW:INT3:
+  description: "An example integer PV with multiple forward links"
+  type: "INTEGER"
+  initial: 0
+  #forward_links: ["DEV:RW:INT1", "DEV:RW:DOUBLECALC"]
+
+DEV:RW:INT4:
+  description: "An example integer PV with multiple forward links"
+  type: "INTEGER"
+  initial: 0
+  #forward_links: ["DEV:RW:INT1", "DEV:RW:DOUBLECALC"]
+
+DEV:RW:DOUBLECALC:
+  description: "A PV with a calc and display limits configured but no precision etc"
+  type: "DOUBLE"
+  initial: 6.0
+  display:
+    low: -1.0
+    high: 1.0
+  calc:
+    calc_str: "pv[0]+2.12*pv[1]"
+    variables: ["DEV:RW:INT3", "DEV:RW:INT4"]
+    #variables: ["DEV:RW:DOUBLE1", "DEV:RW:DOUBLE2"]
+    #variables: ["DEV:RW:INT1", "my:pv:name"]
+    #calc_str: "pv[0]+2.12"
+    #variables: ["my:pv:name"]
+
+DEV:RW:INT:
+  description: "An example int PV with limits"
+  type: "INTEGER"
+  initial: 3
+
+DEV:RW:INT1:
+  description: "An example integer PV"
+  type: "INTEGER"
+  initial: 2
+
+# DEV:RO:INT:
+#   description: "An example int PV with limits"
+#   type: "INTEGER"
+#   initial: 3
+#   # read_only: True
+#   control:
+#     low: 1
+#     high: 10
+#   valueAlarm:
+
+# DEV:RW:STRING1:
+#   description: "An example string PV"
+#   type: "STRING"
+#   initial: 5.0
+
+# DEV:RW:ENUM:
+#   description: "An example enum PV"
+#   type: "ENUM"
+#   initial:
+#     choices: ['OFF', 'ON']
+#     index: 0
