+++ tests/unit/thread/test_config_reader.py
@@ -3,8 +3,8 @@ from unittest.mock import MagicMock
 import numpy as np
 import pytest
 
+from p4pillon.config_reader import parse_config
 from p4pillon.definitions import PVTypes
-from p4pillon.thread.config_reader import parse_config
 from p4pillon.thread.pvrecipe import PVEnumRecipe, PVScalarArrayRecipe, PVScalarRecipe
